Flytt sikkerhets-config ting til egen pakke
Trekk ut prosseseringConfig til egen fil
Bruk restTestClient i stedet for webTestClient
Ikke gjør flyway-endringer som allerede er deployet
