+++ pom.xml
@@ -27,7 +27,7 @@
 
         <felles.version>4.20260428145828_c16a1d8</felles.version>
         <kontrakter.version>4.0_20260427083013_f53660d</kontrakter.version>
-        <mock-oauth2-server.version>2.1.10</mock-oauth2-server.version>
+        <mock-oauth2-server.version>3.0.1</mock-oauth2-server.version>
 
         <main-class>no.nav.familie.baks.mottak.LauncherKt</main-class>
 
@@ -292,11 +292,6 @@
             <groupId>no.nav.security</groupId>
             <artifactId>mock-oauth2-server</artifactId>
             <version>${mock-oauth2-server.version}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-webflux</artifactId>
             <scope>test</scope>
         </dependency>
     </dependencies>
