+++ src/main/resources/db/migration/V2__payload-not-unique.sql
@@ -1,2 +1,2 @@
 alter TABLE task
-    DROP CONSTRAINT IF EXISTS task_payload_key;
+    DROP CONSTRAINT task_payload_key;
