+++ src/test/resources/application-testcontainers.yaml
@@ -1,3 +1,3 @@
 spring:
   datasource:
-    url: jdbc:tc:postgresql:17://localhost/familie-ba-sak
+    url: jdbc:tc:postgresql:17://localhost/familie-baks-mottak
